+++ app/models/transaction.rb
@@ -7,13 +7,8 @@ class Transaction < ApplicationRecord
   belongs_to :report, optional: true
 
   validates_presence_of :report, unless: -> { parent_activity&.organisation&.service_owner? }
-  validates_presence_of :description,
-    :transaction_type,
+  validates_presence_of :value,
     :date,
-    :currency,
-    :value,
-    :providing_organisation_name,
-    :providing_organisation_type,
     :receiving_organisation_name,
     :receiving_organisation_type
   validates :value, numericality: {other_than: 0, less_than_or_equal_to: 99_999_999_999.00}